Fix bug 5087752 - Maintain correct contrast against action bars in inverse-bar themes
Add the actionBarWidgetTheme theme attribute. This lets a theme
specify a wrapper theme that can be used to create views that will
end up in the action bar so that the rest of the code can ignore
differences in contrast. (e.g. the inverse action bar themes.)

Apps can use ActionBar#getThemedContext() to obtain a Context with a
proper theme for views that will end up in the action
bar. MenuInflaters generated by Activities will automatically use this
to properly theme inflated action views.

Change-Id: Ib28c82bf47c25d446cca2a63f617b8a4a0afa6b2