Handle rotation gracefully in the backup/restore confirmation UI
We now don't automatically deny the operation if stopped, but instead
allow the activity to be destroyed and recreated as usual.  We retain
the observer instance across that sequence so we keep getting progress
reports etc.

The UI now also uses the spiffy new button bar styles, and positions
the deny / confirm buttons according to ICS standards.

Bug 5115411

Change-Id: Ie760a0c8496c69f9d5881273a63ad5b5b76ff554