Don't do full backup/restore before setup
On the restore side, there's a bunch of one-time setup, device
provisioning, etc that we're very much not prepared to do in
lieu of running setup wizard, at least at this time.

On the backup side, it simply doesn't make sense to back up
stuff before the device has been set up.

Part of bug 5290261

Change-Id: If1c65e88e2da589d6204232d2b59c3e994f4ed3f