TouchExplorer crashes if there is incative pointer while dragging.
The TouchExplorer was not taking into account the case with incative
pointers while dragging. If one puts a finger down and then perfroms
a dragging gestore the explorer tries to inject UP event for the end
of the gesture upon every of the two dragging pointers going up instead
only for one the first went up.

bug:5476098

Change-Id: I20d2dd7bde7e016b0678a35d14cd068d9ff37023