Fix bug in enabling/disabling SMS cell broadcast activation.
When enabling/disabling SMS cell broadcast channels, we were not calling
setGsmBroadcastActivation() with the correct value after updating the
message IDs. It should be called with true if any message IDs are enabled,
or false if the list is empty.

Added an isEmpty() method to IntRangeManager, and moved the call to
setGsmBroadcastActivation() in SimSmsInterfaceManager to the end of the
enableCellBroadcastRange() and disableCellBroadcastRange() methods,
where it sets the correct value using the new isEmpty() method to test
if there are any message IDs enabled after updating the range list.

Bug: 5525441
Change-Id: I7d1ebd54dacf1de20910947efbf5e87e1957fd1a