Fix device management in lvm effect bundle wrapper.
The virtualizer and bass boost effects should be disabled when
not playing throught headsets or headphones.

There is a bug in current wrapper implementation that makes that those
effects are not disabled if the effect is enabled by the user after
the headset has been removed (current logic only works if the device selection
occurs while the effect is enabled).

Change-Id: I7c66b15a0339d95fb9ed13c8320d66379725d3b6