Fix bug 5594320 - Overlay action bar decor layout causes two measure passes
The existing action bar overlay decor layouts hit a case in
RelativeLayout that causes two measure passes for the main content. As
this can be expensive, stick the bar and the content overlay into
their own sub-layout and switch things to use a FrameLayout at the top
level instead.

Be explicit about the layout_width/height on ActionBar-containing
decor layouts as the older decor layouts are.

Change-Id: I4330f0f25841dd8058b94a320f93bf67fb56bf17