No blackscreen when dialog is showing fixes 5511992
When a dialog is activated, we won't show the blackscreen
until after the lockscreen comes back in focus, which occurs
when the dialog closes.
The user may toggle the power button, causing onScreenTurnedOff to
reset mHasOverlay, even though the dialog is still on top of the window.
Thus, the dialog will have a black box behind it without this change.

Change-Id: I6bec7984b88454e9fff22147cbbfc27b9c0a74f2